Returned some of the original code to Turn2Urn.atl
Tracking which pieces of code break the system in an effort to repair them one at a time. The rules that break the system at present can be found Turn2UrnModifiedToFix.txt. I have also returned the createTimer method to Turn2Urn.atl; it was lost at some point in my refactoring.