Started returning code from ConnectNodes
I am returning the code in segments to test where the inefficiencies are. At present, I am approaching this by dividing the code into multiple "paths" corresponding to different branches. I have commented the name of each path into ConnectNodes and the code for these paths is stored in the "ConnectNodes logic" folder.