Fixed a runtime error caused by a bug in createAndFork
An invalid variable was being used in the method. I resolved this by following the equivalent logic in createOrFork in assuming the two work in a similar manner. I am not certain this is a correct assumption, but will leave determining that for when I am able to produce a more full output file.