Finally started refactoring of ConnectNodes
I will need to fix it to figure out what is going wrong with the system. This will take some time and I will do it in sweeps. At present, I am separating out duplicate logic without really trying to understand what it does. This will reduce the overall amount of code, ideally improving readability enough for me to understand what is happening in the code.