Created a called rule in ConnectNodes
Name is not very helpful - "doThing" - as I cannot determine what the
logic does. Note that the current version of the code is broken and will remain so
until I have condensed the logic in ConnectNodes. Without this condensing, it takes unreasonably long to build. As such, I will be editing the file directly in a separate IDE so that it does not autobuild. This means I will not be able to verify syntax for now.