Wrote a helper, renamed another helper
Wrote the handleConnectNode rule. Although I am not sure precisely what it does, it is only called in the context of Connect nodes. Also renamed doThing to processNode for a meaningful name, even if it is not very specific. Note that this is still a first iteration of refactoring this code, and I expect to optimize the readability further in the next iteration. I did miss a segment of code with a potential handleConnectNode call. I will handle it on a later day.