Added a variable to ConnectNodes. Fixed all compilation errors
Added a thisModule.firstPathNode variable to ConnectNodes for readability. Build times are now more reasonable so I have also been able to resolve all compiler errors. I have not tried executing yet - that is a task for tomorrow. I am pleased to note that building the workspace now takes less than a minute. I expect to improve that further in the time to come. However, I have absolutely no guarantee of the correctness of the system at present - I am working on a major refactoring task without any way to verify the correctness of the system until I am done