Removed many 'do' blocks from matched rules in favor of 'using' blocks
Found that 'using' blocks are a convenient way to define variables
without needing to set them as properties of the module (i.e., it avoid
using thisModule). This should significantly improve readability, as
well as improve adherence to declarative programming principles, so I
will be making many changes in moving towards 'using' blocks.