2 more 'using' blocks, reworked handleConnectNode to be matched
Added a 'using' block to getNodeToConnect and to processNode.
handleConnectNode had creation of the same types of objects in both of
its conditional cases. This meant matched rules were possible.  This is good
as it much more in the spirit of declarative programming. A consequence of
this change is thathe order in which ids are assigned has changed but otherwise
the logic should have the same effect as before.