Updated 2 rules & removed 1, increased README + comment precision
Updated processPathEnd to no longer use the includeBackLinks boolean,
opting to move that logic out of the rule instead. Removed
createAndForkNodeConnection as it was simply equivalent to
createNodeConnection. Updated createOrForkNodeConnection to
no longer take in a tUCMmap as an argument as this was not
being used anyways. Updated comments relating to the third
bug I found in both Turn2Urn and in the README.

Updated readme