Cleanup of retrieveNode, made more calls to it
Removed processTimer logic from retrieveNode as a getter method should
not be modifying the state of the system. Made more calls to
retrieveNode and discovered a bug caused by attempting to get items with
null names from the tUCMmap by name. This is now resolved in some cases
by adding those calls to retrieveNode. Added comment explaining the else
case of retrieveNode,