More retrieveNode calls, commented out some code
Made calls to retrieveNode wherever it did not change the output of the
system to streamline getting nodes. Calls are now made in place of
tUCMmap.getNode in all places where it does not imapct the output. Some
resolveTemps have also been replaced with such calls. Commeneted out
some code that didn't seem to be changing the output of the system in
any tests.