Minor rework of createOrFork, some cleanup
Rework so that predecessors are no longer passed in as an argument.
Having to declare them explicitly elsewhere will make reading the code
more easy. Also replaced another resolveTemp for a retrieveNode and
commented out more code that didn't seem to modify the state of the
system in any test.