More rearranging; removed two rules; wrote a rule
Removed the processPathNode rule as it wasn't impacting the output except in one location.
Moved its logic to that one location, in connectOrForkEndPoints, though
this logic is commented out for now as I have yet to further test it.
Removed collectRegularEnds as almost all of its logic was handled in
processPathEnd.
Wrote the new rule processEndpointWithConnect to handle some logic that was previously in processPathEnd.
Now processPathEnd instead makes a call to this new rule.